% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_multiclass}
\alias{fit_multiclass}
\title{Fit Multiclass Model}
\usage{
fit_multiclass(shap_features, y, alpha, return_individual = FALSE)
}
\arguments{
\item{shap_features}{List of data frames with SHAP values per class}

\item{y}{Target vector with multiple classes}

\item{alpha}{Regularization parameter}

\item{return_individual}{Logical, return individual results}
}
\value{
Data frame with combined significance results
}
\description{
Perform one-vs-all logistic regression for each class
}
\keyword{internal}
