% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shap_select.R
\name{iterative_reduction}
\alias{iterative_reduction}
\title{Iterative Feature Reduction}
\usage{
iterative_reduction(shap_features, y, task, alpha = 1e-06)
}
\arguments{
\item{shap_features}{Data frame(s) with SHAP values}

\item{y}{Target vector}

\item{task}{Character: task type}

\item{alpha}{Regularization parameter}
}
\value{
Data frame with feature elimination sequence
}
\description{
Recursively remove least significant features
}
\keyword{internal}
