% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_shap.R
\name{remove_feature}
\alias{remove_feature}
\title{Remove Feature from SHAP Features}
\usage{
remove_feature(shap_features, feature_name)
}
\arguments{
\item{shap_features}{Data frame or list of data frames}

\item{feature_name}{Character: name of feature to remove}
}
\value{
Data frame or list with feature removed
}
\description{
Remove a feature from data frame or list of data frames
}
\keyword{internal}
