% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsshap.R
\name{shap_features_to_significance}
\alias{shap_features_to_significance}
\title{SHAP Features to Significance}
\usage{
shap_features_to_significance(shap_features, y, task, alpha)
}
\arguments{
\item{shap_features}{Data frame(s) with SHAP values}

\item{y}{Target vector}

\item{task}{Character: "regression", "binary", or "multiclass"}

\item{alpha}{Regularization parameter}
}
\value{
Data frame sorted by t-value (descending)
}
\description{
Determine task type and compute significance for all features
}
\keyword{internal}
